test(ClienteDAOTest): add mais testes

+++ src/main/java/org/es2locadora/cliente/Cliente.java
@@ -21,10 +21,9 @@ public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_cliente")
-    private Integer idCliente;
+    private Long idCliente;
 
     @Column(name = "nome_cliente")
-
     @NonNull
     private String nomeCliente;
